pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qonboard"
version = "1.0.0"
description = "Quilr customer onboarding agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt==4.3.0",
    "rich==13.9.4",
    "openai==1.55.3",
    "jira==3.8.0",
    "requests==2.32.3",
    "psycopg2-binary==2.9.10",
    "neo4j==5.18.0",
    "python-dotenv==1.0.1",
]

[project.scripts]
qonboard = "qonboard.agent:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["qonboard*"]
